점검 페이지 포커스 배너 버전 관리 설정 추가

+++ src/db/CREATE.sql
@@ -23,7 +23,7 @@ DROP TABLE IF EXISTS ref_focus_banner_info;
 CREATE TABLE ref_focus_banner_info (
     id          INT AUTO_INCREMENT PRIMARY KEY COMMENT '기본 키 컬럼',
     version     INT             NOT NULL COMMENT '배너 버전',
-    banner_type TINYINT         NOT NULL COMMENT '배너 타입 (ex. 1: 메인 포커스 배너, 2: 뉴스 포커스 배너 첫번째, 3: 뉴스 포커스 배너 두번째)',
+    banner_type SMALLINT        NOT NULL COMMENT '배너 타입 (ex. 1: 메인 포커스 배너, 2: 뉴스 포커스 배너 첫번째, 3: 뉴스 포커스 배너 두번째)',
     -- TODO banner_url 컬럼을 추가해서 해당 배너와 실제 연결되는 페이지 링크 정보도 DB 에서 관리하도록 하기
     img_url     VARCHAR(255)    NOT NULL COMMENT '이미지 소스 링크',
     img_order   SMALLINT        NOT NULL COMMENT '이미지 출력 순서',
